+++ generator.h
@@ -1,6 +1,6 @@
-int *generateSortedArray(int n);
-int *generateReverseSortedArray(int n);
-int *generateRandomizedArrayRangeN(int n);
-int *generateRandomizedArrayRangeLogN(int n);
-int *generateRandomizedArrayRangeNSquare(int n);
-int *generateConstantArray(int n, int constant);
+long *generateSortedArray(long n);
+long *generateReverseSortedArray(long n);
+long *generateRandomizedArrayRangeN(long n);
+long *generateRandomizedArrayRangeLogN(long n);
+long *generateRandomizedArrayRangeNSquare(long n);
+long *generateConstantArray(long n, long constant);
